% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormy.R
\name{rnormy}
\alias{rnormy}
\title{R final project}
\usage{
rnormy(p, u1, u2, s1, s2)
}
\arguments{
\item{p}{a number}

\item{u1}{a number}

\item{u2}{a number}

\item{s1}{a number}

\item{s2}{a number}
}
\value{
vector y from mixture normal distribution
}
\description{
R final project
}
\keyword{distribution}
\keyword{mixture}
\keyword{normal}
