% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{R final project}
\usage{
weather(x)
}
\arguments{
\item{x}{a number}
}
\value{
the next 10 days weather
}
\description{
R final project
}
\keyword{simulation}
\keyword{weather}
